fix for empty body on 204 status code
According to HTTP spec,status code 204 should not have a body
The current code checks for 'response.body' to achieve this.

However Rack sends empty string as body for a 204 status code which is considered a body when using the response.body check.
As a result deserializing from Json fails. using .present? solves this issue (borrowed rails .present? implementation here as rails will not always be available)

related rack information:
- example use in test : line 207 in https://github.com/rack/rack/blob/master/test/spec_response.rb
- code : line 63 in https://github.com/rack/rack/blob/master/lib/rack/response.rb
- code introduced in commit https://github.com/rack/rack/commit/c525c2871a0de628786233077ac3603e9fe5d8e4
- rack specifications : lines 238 till 244  in https://github.com/rack/rack/blob/master/SPEC